caller_v2/requirements.txt
alembic==1.7.1
docker==5.0.2
fastapi==0.68.1
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
requests==2.26.0
SQLAlchemy==1.4.23
uvicorn==0.15.0
